requirements.txt
# local package
-e .

# external requirements
click
flake8
python-dotenv>=0.5.1
